added DatabaseConfig.maxLifeTimeInMillis - ran a8-codegen this time

+++ shared/shared/src/main/scala/a8/shared/jdbcf/DatabaseConfig.scala
@@ -23,6 +23,6 @@ case class DatabaseConfig(
   password: Password,
   minIdle: Int = 1,
   maxPoolSize: Int = 50,
-  maxLifeTimeInMillis: Option[Int] = None,
+  maxLifeTimeInMillis: Option[Long] = None,
   autoCommit: Boolean = true,
 )
